A hack to work around a bug in eshell
As best as I can explain it, the following steps reproduce the bug

- start eshell
- cd /ssh:<some-host>:
- type `ls /ho` and press TAB to trigger auto completion
- Pick one of the candidates
- Press ENTER to run the command

The command will likely take a few seconds to run.  I think the following is
happening
- the autocompletion of `ls /ho` leaves `completion-in-region-mode` turned
on (the BUG)
- as soon as the command finish running,
completion-in-region--postch (post-command-hook) run trying to autocomplete the
empty string (e.g. `user@10-1.1.1$ `) which ends up scanning the entire PATH
for programs that could be used.

The defadvice ensures the mode is turned off before sending the input to the
subprocess.